Implement half precision for P values

+++ tomobar/regularisersCuPy.py
@@ -118,15 +118,15 @@ def PD_TV_cupy(
 
     # initialise CuPy arrays here:
     U_arrays = [data.copy(), cp.zeros(data.shape, dtype=cp.float32, order="C")]
-    P1_arrays = [cp.zeros(data.shape, dtype=cp.float32, order="C") for _ in range(2)]
-    P2_arrays = [cp.zeros(data.shape, dtype=cp.float32, order="C") for _ in range(2)]
+    P1_arrays = [cp.zeros(data.shape, dtype=cp.float16, order="C") for _ in range(2)]
+    P2_arrays = [cp.zeros(data.shape, dtype=cp.float16, order="C") for _ in range(2)]
 
     # loading and compiling CUDA kernels:
     module = load_cuda_module("primal_dual_for_total_variation")
     if data.ndim == 3:
         data3d = True
         P3_arrays = [
-            cp.zeros(data.shape, dtype=cp.float32, order="C") for _ in range(2)
+            cp.zeros(data.shape, dtype=cp.float16, order="C") for _ in range(2)
         ]
         dz, dy, dx = data.shape
         # setting grid/block parameters
